+++ lib/ui/groups/GroupMemberNotificationsPanel.dart
@@ -443,7 +443,7 @@ class _GroupMemberNotificationsPanelState extends State<GroupMemberNotifications
 
 class _DisabledToggleButton extends ToggleRibbonButton{
   static Map<bool, Widget> _rightIcons= {
-    true: Styles().images.getImage('check-green', excludeFromSemantics: true) ?? Container(),
+    true: Styles().images.getImage('check-accent', excludeFromSemantics: true) ?? Container(),
     false: Styles().images.getImage('close', excludeFromSemantics: true) ?? Container(),
   };